small foolproof protection again deleting wrong files

+++ Sm5sh.CLI/WorkspaceManager.cs
@@ -66,8 +66,15 @@ namespace Sm5sh.CLI
             _logger.LogInformation("Cleaning up workspace folder");
             foreach (var fileToDelete in existingFiles)
             {
-                File.Delete(fileToDelete);
+                if (IsWorkspaceFile(fileToDelete))
+                    File.Delete(fileToDelete);
             }
+        }
+
+        private bool IsWorkspaceFile(string filename)
+        {
+            var file = Path.GetFileName(filename).ToLower();
+            return file.EndsWith("msbt") || file.EndsWith("nus3audio") || file.EndsWith("nus3bank") || file.EndsWith("prc") || file == "bgm_property.bin";
         }
     }
 }
